backend-app-api: remove startHttpServer
